jwt_auth/requirement.txt
asgiref==3.8.1
certifi==2024.8.30
dj-database-url==2.2.0
Django==5.1.2
django-cors-headers==4.4.0
django-jazzmin==3.0.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
